+++ docs/json.rst
@@ -1,0 +1,11 @@
+JSON
+====
+.. automodule:: numcodecs.json
+
+.. autoclass:: JSON
+
+    .. autoattribute:: codec_id
+    .. automethod:: encode
+    .. automethod:: decode
+    .. automethod:: get_config
+    .. automethod:: from_config
